requirements.txt
pandas>=1.4
numpy>=1.21
openai>=1.0.0
tqdm>=4.62
pdf2image>=1.16
PyPDF2>=3.0
Pillow>=9.0
streamlit>=1.20